Add files via upload
Doubly Linked List is a variation of the Linked list in which navigation is possible in both ways, either forward and backward easily as compared to Single Linked List. It has a prev and next and the value. It means that a single node has 3 values where one is Previous address of the node, other is the next address to the node and last is the value of that particular node.