Add files via upload
Doubly Linked List is a variation of the Linked list in which navigation is possible in both ways, either forward and backward easily as compared to Single Linked List. It has a prev and next and the value. It means that a single node has 3 values where one is Previous address of the node, other is the next address to the node and last is the value of that particular node.

But at last, the first element has the address of the last as well as the first in its prev and next. Similarly, for the last node, it stores first node address in the next node and last but one address in the prev.